+++ utils/tools.py
@@ -157,7 +157,7 @@ class PointCloudsInPickle(InMemoryDataset):
 
         sample = Data(
             x=torch.from_numpy(x).float(),
-            y=torch.from_numpy(np.array(target)).type(torch.half),
+            y=torch.from_numpy(np.array(target)).float(),
             pos=torch.from_numpy(coords[use_idx, :]).float(),
         )
         if coords.shape[0] < 100:
